+++ ADotNet/Models/Pipelines/GithubPipelines/DotNets/Tasks/Builders/NuGetPushTaskBuilder.cs
@@ -4,9 +4,7 @@
 // See License.txt in the project root for license information.
 // ---------------------------------------------------------------------------
 
-using ADotNet.Models.Pipelines.GithubPipelines.DotNets.Tasks;
-
-namespace ADotNet.Models.Pipelines.GithubPipelines.DotNets.TaskBuilders
+namespace ADotNet.Models.Pipelines.GithubPipelines.DotNets.Tasks.Builders
 {
     public class NuGetPushTaskBuilder
     {
@@ -14,51 +12,45 @@ namespace ADotNet.Models.Pipelines.GithubPipelines.DotNets.TaskBuilders
 
         public NuGetPushTaskBuilder()
         {
-            this.nugetPushTask = new NuGetPushTask();
+            nugetPushTask = new NuGetPushTask();
         }
 
         public NuGetPushTaskBuilder WithName(string name)
         {
-            this.nugetPushTask.Name = name;
+            nugetPushTask.Name = name;
             return this;
         }
 
         public NuGetPushTaskBuilder WithSearchPath(string searchPath)
         {
-            this.nugetPushTask.SearchPath = searchPath;
+            nugetPushTask.SearchPath = searchPath;
             return this;
         }
 
         public NuGetPushTaskBuilder WithApiKey(string apiKey)
         {
-            this.nugetPushTask.ApiKey = apiKey;
+            nugetPushTask.ApiKey = apiKey;
             return this;
         }
 
         public NuGetPushTaskBuilder WithDestination(string destination)
         {
-            this.nugetPushTask.Destination = destination;
+            nugetPushTask.Destination = destination;
             return this;
         }
 
         public NuGetPushTask Build()
         {
-            if (!string.IsNullOrEmpty(this.nugetPushTask.SearchPath))
-            {
-                this.nugetPushTask.Run += $" \"{this.nugetPushTask.SearchPath}\"";
-            }
+            if (!string.IsNullOrEmpty(nugetPushTask.SearchPath))
+                nugetPushTask.Run += $" \"{nugetPushTask.SearchPath}\"";
 
-            if (!string.IsNullOrEmpty(this.nugetPushTask.ApiKey))
-            {
-                this.nugetPushTask.Run += $" --api-key {this.nugetPushTask.ApiKey}";
-            }
+            if (!string.IsNullOrEmpty(nugetPushTask.ApiKey))
+                nugetPushTask.Run += $" --api-key {nugetPushTask.ApiKey}";
 
-            if (!string.IsNullOrEmpty(this.nugetPushTask.Destination))
-            {
-                this.nugetPushTask.Run += $" --source \"{this.nugetPushTask.Destination}\"";
-            }
+            if (!string.IsNullOrEmpty(nugetPushTask.Destination))
+                nugetPushTask.Run += $" --source \"{nugetPushTask.Destination}\"";
 
-            return this.nugetPushTask;
+            return nugetPushTask;
         }
     }
 }
